clean up of repo and restore of userController.js
