+++ logs/access.log
@@ -1,58 +1,2 @@
-::1 - - [01/Sep/2024:11:13:21 +0000] "GET /status HTTP/1.1" 404 145 "-" "curl/8.5.0"
-::1 - - [01/Sep/2024:11:13:34 +0000] "GET /api/status HTTP/1.1" 200 11 "-" "curl/8.5.0"
-::1 - - [01/Sep/2024:11:20:49 +0000] "GET /api/status HTTP/1.1" 200 42 "-" "curl/8.5.0"
-::1 - - [01/Sep/2024:11:22:00 +0000] "GET /api/status HTTP/1.1" 200 38 "-" "curl/8.5.0"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:12 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:13 +0000] "GET /api/stats HTTP/1.1" 200 23 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:18 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:18 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:18 +0000] "POST /api/journal-entries HTTP/1.1" 201 244 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "GET /api/stats HTTP/1.1" 200 23 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "POST /api/user/register HTTP/1.1" 400 31 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:05:37 +0000] "POST /api/journal-entries HTTP/1.1" 201 244 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:07:59 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:07:59 +0000] "GET /api/stats HTTP/1.1" 200 23 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:07:59 +0000] "POST /api/user/register HTTP/1.1" 400 31 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:07:59 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:08:00 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:21:08:00 +0000] "POST /api/journal-entries HTTP/1.1" 201 244 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:20 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:21 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:21 +0000] "POST /api/user/register HTTP/1.1" 400 31 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:21 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:21 +0000] "POST /api/user/login HTTP/1.1" 200 285 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:37:21 +0000] "POST /api/journal-entries HTTP/1.1" 201 244 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:49:10 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:49:10 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:49:10 +0000] "POST /api/journal-entries HTTP/1.1" 401 27 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:49:10 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:52:19 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:52:19 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:52:19 +0000] "POST /api/journal-entries HTTP/1.1" 401 27 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:52:19 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "POST /api/auth/request-password-reset HTTP/1.1" 404 28 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "POST /api/journal-entries HTTP/1.1" 401 27 "-" "-"
-::ffff:127.0.0.1 - - [02/Sep/2024:23:54:46 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "POST /api/auth/request-password-reset HTTP/1.1" 404 28 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "POST /api/journal-entries HTTP/1.1" 401 27 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:05:35 +0000] "GET /api/user/66d212bb0ec12d0c27b00064/journal-entries HTTP/1.1" 200 54 "-" "-"
-::1 - - [03/Sep/2024:00:07:00 +0000] "GET /api/user/66d212bb0ec12d0c27b00064/journal-entries HTTP/1.1" 200 54 "-" "curl/8.5.0"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "GET /api/status HTTP/1.1" 200 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "POST /api/journal-entries HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "POST /api/auth/request-password-reset HTTP/1.1" 404 28 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "GET /api/stats HTTP/1.1" 200 24 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:07:34 +0000] "GET /api/user/66d212bb0ec12d0c27b00064/journal-entries HTTP/1.1" 200 54 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:12:01 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
-::ffff:127.0.0.1 - - [03/Sep/2024:00:12:01 +0000] "POST /api/auth/request-password-reset HTTP/1.1" 404 28 "-" "-"
+::ffff:127.0.0.1 - - [03/Sep/2024:08:18:28 +0000] "POST /api/user/login HTTP/1.1" 500 35 "-" "-"
+::ffff:127.0.0.1 - - [03/Sep/2024:08:18:28 +0000] "POST /api/journal-entries HTTP/1.1" 401 27 "-" "-"
