+++ src/controllers/userController.js
@@ -1,266 +1,184 @@
-import bcrypt from 'bcrypt';
-import request from 'supertest';
-import express from 'express';
-import { registerUser, getUserProfile, updateUserProfile, deleteUserAccount, updateUserPassword } from '../../src/controllers/userController';
-import User from '../../src/models/User';
-import JournalEntry from '../../src/models/JournalEntry';
-import { extractToken, blacklistToken, extractTokenExpiration } from '../../src/utils/jwt';
-import { sendWelcomeMail, sendProfileUpdatedMail, sendPasswordChangedMail, sendAccountDeletedMail } from '../../src/utils/mailer';
-import { logger } from '../../src/middleware/logger';
-import router from '../../src/routes/index';
+import bcrypt from 'bcrypt'; // Library for hashing passwords
+import User from '../models/User'; // User model for interacting with users
+import { extractToken, blacklistToken, extractTokenExpiration } from '../utils/jwt'; // Utility function to extract JWT from request headers
+import JournalEntry from '../models/JournalEntry';
+import { sendWelcomeMail, sendProfileUpdatedMail, sendPasswordChangedMail, sendAccountDeletedMail } from '../utils/mailer';
+import {
+  HTTP_STATUS_OK,
+  HTTP_STATUS_CREATED,
+  HTTP_STATUS_BAD_REQUEST,
+  HTTP_STATUS_UNAUTHORIZED,
+  HTTP_STATUS_NOT_FOUND,
+  HTTP_STATUS_INTERNAL_SERVER_ERROR,
+} from '../constants/httpStatusCodes';
+import dotenv from 'dotenv';
+import { logger } from '../middleware/logger'; // Import the logger
+dotenv.config();
 
-const app = express();
-app.use(express.json());
-app.use(router);
+const saltRounds = parseInt(process.env.SALT_ROUNDS);
+if (!saltRounds) {
+  throw new Error('SALT_ROUNDS environment variable is not set');
+}
 
-jest.mock('../../src/models/User');
-jest.mock('../../src/models/JournalEntry');
-jest.mock('bcrypt');
-jest.mock('../../src/utils/jwt');
-jest.mock('../../src/utils/mailer');
-jest.mock('../../src/middleware/logger');
+/**
+ * Registers a new user.
+ * @param {Object} req - The request object.
+ * @param {Object} res - The response object.
+ * @returns {Promise<Object>} The response containing the new user ID or an error message.
+ */
+async function registerUser(req, res) {
+  const { fullName, nickname, email, password } = req.body; // Extract user data from request body
+  try {
+    // Check if user already exists
+    const existingUser = await User.findByEmail(email);
+    if (existingUser) {
+      logger.info(`Registration failed: User already exists (email: ${email})`, { statusCode: HTTP_STATUS_BAD_REQUEST });
+      return res.status(HTTP_STATUS_BAD_REQUEST).json({ error: 'User already exists' });
+    }
 
-describe('UserController', () => {
-  let req, res;
+    // Hash the password
+    const hashedPassword = await bcrypt.hash(password, saltRounds);
 
-  beforeEach(() => {
-    req = { body: {}, params: {}, headers: {} };
-    res = {
-      status: jest.fn().mockReturnThis(),
-      json: jest.fn(),
-    };
-  });
+    // Create new user instance
+    const newUser = new User({ fullName, nickname, email, password: hashedPassword });
+    const userId = await newUser.save(); // Save new user to the database
 
-  afterEach(() => {
-    jest.clearAllMocks(); // Clears all mocks after each test
-  });
+    // Send welcome email
+    await sendWelcomeMail(email);
+    logger.info(`New user registered successfully (userId: ${userId}, email: ${email})`, { statusCode: HTTP_STATUS_CREATED });
 
-  describe('registerUser', () => {
-    it('should return an error if the user already exists', async () => {
-      User.findByEmail.mockResolvedValue(true);
+    return res.status(HTTP_STATUS_CREATED).json({ userId }); // Respond with the newly created user's ID
+  } catch (error) {
+    logger.error(`Registration error: ${error.message}`, { statusCode: HTTP_STATUS_INTERNAL_SERVER_ERROR });
+    return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).json({ error: 'Error Registering new user' }); // Respond with a server error status
+  }
+}
 
-      req.body = { fullName: 'John Doe', nickname: 'johndoe', email: 'john@example.com', password: 'password123' };
+/**
+ * Fetches a user profile by ID.
+ * @param {Object} req - The request object.
+ * @param {Object} res - The response object.
+ * @returns {Promise<Object>} The response containing the user profile or an error message.
+ */
+async function getUserProfile(req, res) {
+  const userId = req.user.userId; // Extract userId from authenticated user
+  try {
+    const user = await User.findById(userId); // Find user by ID in the database
+    if (!user) {
+      logger.info(`User profile fetch failed: User not found (userId: ${userId})`, { statusCode: HTTP_STATUS_NOT_FOUND });
+      return res.status(HTTP_STATUS_NOT_FOUND).json({ error: 'User not found' });
+    }
 
-      await registerUser(req, res);
+    logger.info(`User profile fetched successfully (userId: ${userId})`, { statusCode: HTTP_STATUS_OK });
+    return res.status(HTTP_STATUS_OK).json({ user }); // Respond with the fetched user profile
+  } catch (error) {
+    logger.error(`Error fetching user profile: ${error.message}`, { statusCode: HTTP_STATUS_INTERNAL_SERVER_ERROR });
+    return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).json({ error: 'Server error' }); // Respond with a server error status
+  }
+}
 
-      expect(res.status).toHaveBeenCalledWith(400);
-      expect(res.json).toHaveBeenCalledWith({ error: 'User already exists' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User already exists'), expect.any(Object));
-    });
+/**
+ * Updates a user profile by ID.
+ * @param {Object} req - The request object.
+ * @param {Object} res - The response object.
+ * @returns {Promise<Object>} The response indicating success or an error message.
+ */
+async function updateUserProfile(req, res) {
+  const userId = req.user.userId; // Extract userId from authenticated user
+  const newData = req.body; // Extract updated data from request body
+  try {
+    const updated = await User.update(userId, newData); // Update user data in the database
+    if (!updated) {
+      logger.info(`User profile update failed: User not found or no changes applied (userId: ${userId})`, { statusCode: HTTP_STATUS_NOT_FOUND });
+      return res.status(HTTP_STATUS_NOT_FOUND).json({ error: 'User not found or no changes applied' });
+    }
+    sendProfileUpdatedMail((newData.email) ? newData.email : req.user.email); // Send profile updated email
+    logger.info(`User profile updated successfully (userId: ${userId})`, { statusCode: HTTP_STATUS_OK });
 
-    it('should create a new user and return the userId', async () => {
-      User.findByEmail.mockResolvedValue(false);
-      bcrypt.hash.mockResolvedValue('hashedpassword');
-      User.prototype.save.mockResolvedValue('newUserId');
-      sendWelcomeMail.mockResolvedValue();
+    return res.status(HTTP_STATUS_OK).json({ message: 'User profile updated successfully' }); // Respond with success message
+  } catch (error) {
+    logger.error(`User profile update error: ${error.message}`, { statusCode: HTTP_STATUS_INTERNAL_SERVER_ERROR });
+    return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).json({ error: 'Server error' }); // Respond with a server error status
+  }
+}
 
-      req.body = { fullName: 'John Doe', nickname: 'johndoe', email: 'john@example.com', password: 'password123' };
+/**
+ * Deletes a user account by ID.
+ * @param {Object} req - The request object.
+ * @param {Object} res - The response object.
+ * @returns {Promise<Object>} The response indicating success or an error message.
+ */
+async function deleteUserAccount(req, res) {
+  const userId = req.user.userId; // Extract userId from authenticated user
+  const email = req.user.email;
+  const token = extractToken(req.headers); // Extract token from request headers
+  try {
+    // Use the JournalEntry class to delete journal entries
+    const stat = await JournalEntry.deleteJournalEntriesByUser(userId); // delete all users journal entries
+    if (!stat) {
+      logger.info(`Delete user account failed: User not found or no journal entries to delete (userId: ${userId})`, { statusCode: HTTP_STATUS_NOT_FOUND });
+      return res.status(HTTP_STATUS_NOT_FOUND).json({ error: 'User not found or no journal entries to delete' });
+    }
+    await User.delete(userId); // Delete user from the database
 
-      await registerUser(req, res);
+    const expirationDate = await extractTokenExpiration(token); // Extract token expiration date
 
-      expect(res.status).toHaveBeenCalledWith(201);
-      expect(res.json).toHaveBeenCalledWith({ userId: 'newUserId' });
-      expect(sendWelcomeMail).toHaveBeenCalledWith('john@example.com');
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('New user registered successfully'), expect.any(Object));
-    });
+    if (!expirationDate) {
+      logger.info('Delete user account failed: Invalid token', { statusCode: HTTP_STATUS_BAD_REQUEST });
+      return res.status(HTTP_STATUS_BAD_REQUEST).send({ error: 'Invalid Token' });
+    }
+    await blacklistToken(token, expirationDate); // Blacklist the token used for authentication
+    sendAccountDeletedMail(email);
+    logger.info(`User account deleted successfully (userId: ${userId})`, { statusCode: HTTP_STATUS_OK });
 
-    it('should handle errors during registration', async () => {
-      User.findByEmail.mockRejectedValue(new Error('Registration error'));
+    return res.status(HTTP_STATUS_OK).json({ message: 'User account deleted successfully' }); // Respond with success message
+  } catch (error) {
+    logger.error(`Delete user account error: ${error.message}`, { statusCode: HTTP_STATUS_INTERNAL_SERVER_ERROR });
+    return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).json({ error: 'Server error' }); // Respond with a server error status
+  }
+}
 
-      req.body = { fullName: 'John Doe', nickname: 'johndoe', email: 'john@example.com', password: 'password123' };
+/**
+ * Updates a user's password.
+ * @param {Object} req - The request object.
+ * @param {Object} res - The response object.
+ * @returns {Promise<Object>} The response indicating success or an error message.
+ */
+async function updateUserPassword(req, res) {
+  const userId = req.user.userId;
+  const { password, newPassword } = req.body;
+  try {
+    const user = await User.findById(userId);
+    if (!user) {
+      logger.info(`Update password failed: User not found (userId: ${userId})`, { statusCode: HTTP_STATUS_NOT_FOUND });
+      return res.status(HTTP_STATUS_NOT_FOUND).json({ error: 'User not found' });
+    }
 
-      await registerUser(req, res);
+    const isPasswordValid = await bcrypt.compare(password, user.password);
+    if (!isPasswordValid) {
+      logger.info(`Update password failed: Invalid password (userId: ${userId})`, { statusCode: HTTP_STATUS_UNAUTHORIZED });
+      return res.status(HTTP_STATUS_UNAUTHORIZED).json({ error: 'Invalid password' });
+    }
 
-      expect(res.status).toHaveBeenCalledWith(500);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Error Registering new user' });
-      expect(logger.error).toHaveBeenCalledWith(expect.stringContaining('Registration error'), expect.any(Object));
-    });
-  });
+    const hashedPassword = await bcrypt.hash(newPassword, saltRounds);
 
-  describe('getUserProfile', () => {
-    it('should return the user profile if the user exists', async () => {
-      req.user = { userId: 'userId' };
-      User.findById.mockResolvedValue({ fullName: 'John Doe', email: 'john@example.com' });
+    await User.update(userId, { password: hashedPassword });
+    sendPasswordChangedMail(user.email);
+    logger.info(`User password updated successfully (userId: ${userId})`, { statusCode: HTTP_STATUS_OK });
 
-      await getUserProfile(req, res);
+    return res.status(HTTP_STATUS_OK).json({ message: 'Password updated successfully' });
 
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.json).toHaveBeenCalledWith({ user: { fullName: 'John Doe', email: 'john@example.com' } });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User profile fetched successfully'), expect.any(Object));
-    });
+  } catch (error) {
+    logger.error(`Update password error: ${error.message}`, { statusCode: HTTP_STATUS_INTERNAL_SERVER_ERROR });
+    return res.status(HTTP_STATUS_INTERNAL_SERVER_ERROR).json({ error: 'Server error' });
+  }
+}
 
-    it('should return an error if the user is not found', async () => {
-      req.user = { userId: 'userId' };
-      User.findById.mockResolvedValue(null);
-
-      await getUserProfile(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(404);
-      expect(res.json).toHaveBeenCalledWith({ error: 'User not found' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User profile fetch failed'), expect.any(Object));
-    });
-
-    it('should handle errors during fetching user profile', async () => {
-      req.user = { userId: 'userId' };
-      User.findById.mockRejectedValue(new Error('Fetch error'));
-
-      await getUserProfile(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(500);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Server error' });
-      expect(logger.error).toHaveBeenCalledWith(expect.stringContaining('Error fetching user profile'), expect.any(Object));
-    });
-  });
-
-  describe('updateUserProfile', () => {
-    it('should update the user profile successfully', async () => {
-      req.user = { userId: 'userId', email: 'old@example.com' };
-      req.body = { fullName: 'John Smith', email: 'new@example.com' };
-      User.update.mockResolvedValue(true);
-      sendProfileUpdatedMail.mockResolvedValue();
-
-      await updateUserProfile(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.json).toHaveBeenCalledWith({ message: 'User profile updated successfully' });
-      expect(sendProfileUpdatedMail).toHaveBeenCalledWith('new@example.com');
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User profile updated successfully'), expect.any(Object));
-    });
-
-    it('should return an error if no changes were applied', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { fullName: 'John Smith' };
-      User.update.mockResolvedValue(false);
-
-      await updateUserProfile(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(404);
-      expect(res.json).toHaveBeenCalledWith({ error: 'User not found or no changes applied' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User profile update failed'), expect.any(Object));
-    });
-
-    it('should handle errors during updating the user profile', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { fullName: 'John Smith' };
-      User.update.mockRejectedValue(new Error('Update error'));
-
-      await updateUserProfile(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(500);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Server error' });
-      expect(logger.error).toHaveBeenCalledWith(expect.stringContaining('User profile update error'), expect.any(Object));
-    });
-  });
-
-  describe('deleteUserAccount', () => {
-    it('should delete the user account successfully', async () => {
-      req.user = { userId: 'userId', email: 'john@example.com' };
-      req.headers.authorization = 'Bearer validToken';
-      JournalEntry.deleteJournalEntriesByUser.mockResolvedValue(true);
-      User.delete.mockResolvedValue(true);
-      extractToken.mockReturnValue('validToken');
-      extractTokenExpiration.mockResolvedValue(new Date());
-      blacklistToken.mockResolvedValue();
-
-      await deleteUserAccount(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.json).toHaveBeenCalledWith({ message: 'User account deleted successfully' });
-      expect(sendAccountDeletedMail).toHaveBeenCalledWith('john@example.com');
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User account deleted successfully'), expect.any(Object));
-    });
-
-    it('should return an error if no journal entries to delete', async () => {
-      req.user = { userId: 'userId' };
-      JournalEntry.deleteJournalEntriesByUser.mockResolvedValue(false);
-
-      await deleteUserAccount(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(404);
-      expect(res.json).toHaveBeenCalledWith({ error: 'User not found or no journal entries to delete' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('Delete user account failed'), expect.any(Object));
-    });
-
-    it('should handle errors during deleting the user account', async () => {
-      req.user = { userId: 'userId' };
-      JournalEntry.deleteJournalEntriesByUser.mockRejectedValue(new Error('Delete error'));
-
-      await deleteUserAccount(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(500);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Server error' });
-      expect(logger.error).toHaveBeenCalledWith(expect.stringContaining('Delete user account error'), expect.any(Object));
-    });
-
-    it('should return an error if the token is invalid', async () => {
-      req.user = { userId: 'userId' };
-      req.headers.authorization = 'Bearer invalidToken';
-      JournalEntry.deleteJournalEntriesByUser.mockResolvedValue(true);
-      User.delete.mockResolvedValue(true);
-      extractToken.mockReturnValue('invalidToken');
-      extractTokenExpiration.mockResolvedValue(null); // Simulate invalid token
-
-      await deleteUserAccount(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(400);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Invalid Token' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('Delete user account failed: Invalid token'), expect.any(Object));
-    });
-  });
-
-  describe('updateUserPassword', () => {
-    it('should update the user password successfully', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { password: 'oldPassword', newPassword: 'newPassword123' };
-      User.findById.mockResolvedValue({ password: 'hashedOldPassword', email: 'john@example.com' });
-      bcrypt.compare.mockResolvedValue(true); // Old password matches
-      bcrypt.hash.mockResolvedValue('hashedNewPassword');
-      User.update.mockResolvedValue(true);
-      sendPasswordChangedMail.mockResolvedValue();
-
-      await updateUserPassword(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.json).toHaveBeenCalledWith({ message: 'Password updated successfully' });
-      expect(sendPasswordChangedMail).toHaveBeenCalledWith('john@example.com');
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('User password updated successfully'), expect.any(Object));
-    });
-
-    it('should return an error if the old password is incorrect', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { password: 'wrongOldPassword', newPassword: 'newPassword123' };
-      User.findById.mockResolvedValue({ password: 'hashedOldPassword', email: 'john@example.com' });
-      bcrypt.compare.mockResolvedValue(false); // Old password does not match
-
-      await updateUserPassword(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(401);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Invalid password' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('Update password failed: Invalid password'), expect.any(Object));
-    });
-
-    it('should return an error if the user is not found', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { password: 'oldPassword', newPassword: 'newPassword123' };
-      User.findById.mockResolvedValue(null); // User not found
-
-      await updateUserPassword(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(404);
-      expect(res.json).toHaveBeenCalledWith({ error: 'User not found' });
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('Update password failed: User not found'), expect.any(Object));
-    });
-
-    it('should handle errors during password update', async () => {
-      req.user = { userId: 'userId' };
-      req.body = { password: 'oldPassword', newPassword: 'newPassword123' };
-      User.findById.mockRejectedValue(new Error('Update password error'));
-
-      await updateUserPassword(req, res);
-
-      expect(res.status).toHaveBeenCalledWith(500);
-      expect(res.json).toHaveBeenCalledWith({ error: 'Server error' });
-      expect(logger.error).toHaveBeenCalledWith(expect.stringContaining('Update password error'), expect.any(Object));
-    });
-  });
-});
+// Export all controller functions as named exports
+export {
+  registerUser,
+  getUserProfile,
+  updateUserProfile,
+  deleteUserAccount,
+  updateUserPassword,
+};
